// src/server/api/models/investing/models/MoneyMarketTransactionRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing a money market transaction request.
 */
export type MoneyMarketTransactionRequest = {
    /**
     * The investment account ID
     */
    accountId: string;
    /**
     * The product ID
     */
    productId: string;
    /**
     * Notes related to the record
     */
    notes?: string;
    /**
     * The transaction currency
     */
    currency: string;
    /**
     * The face value amount
     */
    faceValue: number;
    /**
     * The start date
     */
    startDate: string;
    /**
     * The tenor
     */
    tenor?: number;
    /**
     * The discount rate
     */
    discountRate?: number;
    /**
     * The reference transaction code. This is usually a code generated by an upstream or used for uploading records
     */
    refCode?: string;
    /**
     * The position transfer date is used to indicate that the transaction was migrated from another system and accruals before that date should not be considered.
     */
    positionTransferDate?: string;
    /**
     * The product interest type
     */
    interestType?: MoneyMarketTransactionRequest.interestType;
    /**
     * The channel through which the transaction was submitted
     */
    channel?: MoneyMarketTransactionRequest.channel;
};

export namespace MoneyMarketTransactionRequest {

    /**
     * The product interest type
     */
    export enum interestType {
        CAPITALIZED = 'CAPITALIZED',
        UPFRONT = 'UPFRONT',
    }

    /**
     * The channel through which the transaction was submitted
     */
    export enum channel {
        NA = 'NA',
        WEB = 'WEB',
        MOBILE = 'MOBILE',
        USSD = 'USSD',
        ATM = 'ATM',
        POS = 'POS',
        BRANCH = 'BRANCH',
        EXEC_REPORT = 'EXEC_REPORT',
        FILE_UPLOAD = 'FILE_UPLOAD',
        API = 'API',
        EMAIL = 'EMAIL',
        SMS = 'SMS',
        IVR = 'IVR',
        KIOSK = 'KIOSK',
        CALL_CENTER = 'CALL_CENTER',
        OTHER = 'OTHER',
    }


}